consolidate request
